[ALPS09180239] usb: fix KE because of data type maybe wrong
[Description]
change data type to size_t to avoid KE issue

[Test]
KE not happened when running testcases

MTK-Commit-Id: 4233170ce3a04d30ad5a95df97ace022d397cbe6

CR-Id: ALPS09180239
Feature: usb_musb

+++ drivers/misc/mediatek/usb/usb20/musb_gadget.c
@@ -167,7 +167,7 @@ static inline void map_dma_buffer(struct musb_request *request,
 	struct dma_controller *dma = musb->dma_controller;
 #endif
 
-	unsigned int length;
+	size_t length;
 
 	if (request->request.length == 0)
 		return;
@@ -222,7 +222,7 @@ static inline void map_dma_buffer(struct musb_request *request,
 static inline void
 unmap_dma_buffer(struct musb_request *request, struct musb *musb)
 {
-	unsigned int length;
+	size_t length;
 
 	if (request->request.length == 0)
 		return;
